canary connect

+++ src/modules/purge.ts
@@ -21,19 +21,31 @@ export async function purge(ctx) {
   let langs = await getLang(ctx)
   try {
     if (!ctx.message.reply_to_message) return replyToMessage(ctx, langs.mustReply)
+    if (!await isAdmin(ctx)) {
+      return replyToMessage(ctx, langs.userNonAdmin)
+    }
+    let split = ctx.message.text.split(" ")
+    let silent = split[1] || false
     let now = ctx.message.message_id
     let yesterday = ctx.message.reply_to_message.message_id
     let abs = Math.abs(now - yesterday)
+    let failed = 0
+    ctx.deleteMessage(ctx.message.message_id)
     for (let i = 0; i < abs; i++) {
       jobs.create({
         chat_id: ctx.chat.id,
         message_id: yesterday++
       }, (now, list)=> {
         ctx.telegram.deleteMessage(now.chat_id, now.message_id).catch((e)=> {
+          failed ++
           return e
         })
       })
     }
+    if (!silent) {
+      return replyToMessage(ctx, langs.purgeSuccess.replace(/\{success\}/gmi, Math.floor(abs - failed)).replace(/\{failed\}/gmi, failed))
+    }
+    return
   }catch(error) {
     return reportError(error, ctx)
   }
